Deprecate old keccak (fix #2117)
