--- stdlib/docs/crypto/hashes/keccak_precompile.md
@@ -1,9 +1,0 @@
-
-## std::crypto::hashes::keccak_precompile
-| Procedure | Description |
-| ----------- | ------------- |
-| hash_mem | Computes Keccak256 hash of memory data.<br /><br />Input: [ptr, len_bytes, ...]<br />Output: [keccak_lo, keccak_hi, ...]<br /><br />Where:<br />- ptr: starting memory address (must be word-aligned)<br />- len_bytes: number of bytes to hash<br />- keccak_lo, keccak_hi: 256-bit digest as two words<br /> |
-| hash_mem_impl | Internal implementation of memory-based Keccak256 computation via deferred computation.<br /><br />Emits event to populate advice stack with digest, computes commitment for tracking<br />deferred computations, and stores witness in advice map for later proof generation.<br /><br />Input: [ptr, len_bytes, ...]<br />Output: [commitment, keccak_lo, keccak_hi, ...]<br /><br />Where:<br />- ptr: starting memory address (must be word-aligned)<br />- len_bytes: number of bytes to hash<br />- commitment: RPO(RPO(input) \|\| RPO(hash)) for kernel tracking<br />- keccak_lo, keccak_hi: 256-bit digest as two words<br /> |
-| merge_stack | Merges two 256-bit digests via Keccak256 hash.<br /><br />Input: [digest_left_lo, digest_left_hi, digest_right_lo, digest_right_hi, ...]<br />Output: [keccak_lo, keccak_hi, ...]<br /><br />Where:<br />- digest_left_lo/hi, digest_right_lo/hi: two digests to merge (as four words)<br />- keccak_lo, keccak_hi: Keccak256(left \|\| right) as two words<br /> |
-| merge_stack_impl | Internal implementation of digest merging via deferred Keccak256 computation.<br /><br />Emits event to populate advice stack with digest, computes commitment for tracking<br />deferred computations, and stores witness in advice map for later proof generation.<br /><br />Input: [digest_left_lo, digest_left_hi, digest_right_lo, digest_right_hi, ...]<br />Output: [commitment, keccak_lo, keccak_hi, ...]<br /><br />Where:<br />- digest_left_lo/hi, digest_right_lo/hi: two digests to merge (as four words)<br />- commitment: RPO(RPO(input) \|\| RPO(hash)) for kernel tracking<br />- keccak_lo, keccak_hi: Keccak256(left \|\| right) as two words<br /> |
-| build_commitment_from_digest | Common helper that builds commitment from RPO(input) and advice stack digest.<br /><br />Input: [RPO(input), ...]<br />Output: [commitment, keccak_lo, keccak_hi, ...]<br /><br />Where commitment = RPO(RPO(input) \|\| RPO(keccak))<br /> |
