+++ stdlib/tests/crypto/mod.rs
@@ -7,6 +7,7 @@ mod ecdsa_secp256k1;
 mod elgamal;
 mod fri;
 mod keccak256;
+mod keccak256_deprecated;
 mod rpo;
 mod sha256;
 mod stark;
