--- stdlib/tests/handlers/mod.rs
@@ -1,1 +1,0 @@
-mod keccak;
