+++ stdlib/tests/main.rs
@@ -28,7 +28,6 @@ macro_rules! build_debug_test {
 
 mod collections;
 mod crypto;
-mod handlers;
 mod mast_forest_merge;
 mod math;
 mod mem;
